+++ crates/splitter-cli/src/commands/send.rs
@@ -15,7 +15,7 @@ pub(crate) async fn run(
     input: &str,
     addr: &str,
     stream_id: u8,
-    bitrate: i32,
+    bitrate: u32,
     source: crate::Source,
     fec_mode: crate::SendFecMode,
     simulated_loss_pct: u8,
@@ -32,7 +32,7 @@ pub(crate) async fn run(
     let core_fec_mode = map_fec_mode(fec_mode);
     let mut fec = splitter_core::net::fec::FecController::new(core_fec_mode, 1, 0, 10);
 
-    let mut encoder = OpusEncoder::new(bitrate)?;
+    let mut encoder = OpusEncoder::new(bitrate as i32)?;
     let mut payload_buf = BytesMut::with_capacity(400);
     let mut out_buf = BytesMut::with_capacity(512);
     let mut frame = vec![0.0f32; FRAME_STEREO_SAMPLES];
